refactored catalog classes
split out into public and impl classes cleanly so that connector v2 implementation can use it
