+++ core/src/main/scala/io/snappydata/sql/catalog/SnappyExternalCatalog.scala
@@ -28,10 +28,9 @@ import com.pivotal.gemfirexd.Attribute
 import com.pivotal.gemfirexd.internal.engine.diag.SysVTIs
 import com.pivotal.gemfirexd.internal.iapi.sql.dictionary.SchemaDescriptor
 import com.pivotal.gemfirexd.internal.shared.common.reference.SQLState
+import io.snappydata.Constant
 import io.snappydata.sql.catalog.SnappyExternalCatalog._
-import io.snappydata.{Constant, Property}
 
-import org.apache.spark.SparkEnv
 import org.apache.spark.jdbc.{ConnectionConf, ConnectionUtil}
 import org.apache.spark.sql.catalyst.catalog.{CatalogDatabase, CatalogStorageFormat, CatalogTable, CatalogTableType, ExternalCatalog}
 import org.apache.spark.sql.catalyst.util.CaseInsensitiveMap
@@ -230,10 +229,10 @@ trait SnappyExternalCatalog extends ExternalCatalog {
       case None =>
         val params = new CaseInsensitiveMap(tableDefinition.storage.properties)
         params.get(BASETABLE_PROPERTY) match {
-        // older released didn't have base table entry for indexes
-        case None => params.get(INDEXED_TABLE).map(Utils.toUpperCase)
-        case Some(t) => Some(Utils.toUpperCase(t))
-      }
+          // older released didn't have base table entry for indexes
+          case None => params.get(INDEXED_TABLE).map(Utils.toUpperCase)
+          case Some(t) => Some(Utils.toUpperCase(t))
+        }
       case Some(t) => Some(Utils.toUpperCase(t))
     }
   }
@@ -289,13 +288,6 @@ object SnappyExternalCatalog {
   private[sql] val PASSWORD_MATCH = "(?i)(password|passwd).*".r
 
   val currentFunctionIdentifier = new ThreadLocal[FunctionIdentifier]
-
-  def cacheSize: Int = {
-    SparkEnv.get match {
-      case null => Property.CatalogCacheSize.defaultValue.get
-      case env => Property.CatalogCacheSize.get(env.conf)
-    }
-  }
 
   def getDependents(properties: Map[String, String]): Array[String] = {
     properties.get(DEPENDENT_RELATIONS) match {
