+++ core/src/main/scala/io/snappydata/sql/catalog/impl/StoreHiveCatalog.scala
@@ -14,7 +14,7 @@
  * permissions and limitations under the License. See accompanying
  * LICENSE file.
  */
-package io.snappydata.sql.catalog
+package io.snappydata.sql.catalog.impl
 
 import java.sql.SQLException
 import java.util.Collections
@@ -36,6 +36,7 @@ import com.pivotal.gemfirexd.internal.impl.sql.catalog.GfxdDataDictionary
 import com.pivotal.gemfirexd.internal.shared.common.reference.SQLState
 import io.snappydata.Constant
 import io.snappydata.sql.catalog.SnappyExternalCatalog.checkSchemaPermission
+import io.snappydata.sql.catalog.{CatalogObjectType, ConnectorExternalCatalog, SnappyExternalCatalog}
 import io.snappydata.thrift._
 import org.apache.log4j.{Level, LogManager}
